Some refactoring + ft. add using nested fragments.
There was issue earlier that was not able to access the "_meta" object, which was made as seperate fragment which was included on the navigation query itself.

Now managed to find solution how to use those fragments the "right way" with all the types included. Managed to clean some code way because of this change.